requirements.txt
python-dotenv
imaplib2
pyodbc
langchain
langchain-groq
